style: fix C4244/C4267 in NoteField/TimingSegments/RageSurfaceUtils/ScreenOptionsMasterPrefs (item 2)
Continuing the item-2 sweep (297 -> 187 remaining unique sites). All
casts are documentation of existing intentional behavior, not logic
changes -- characterization-verified (sm_tests unchanged at 5966/226).

- NoteField.cpp (35 call-site warnings, ~30 from two macros):
  * draw_all_segments's `side_sign= cond ? -1 : 1` (int ternary into a
    float) -> `-1.f : 1.f`.
  * IS_ON_SCREEN's two float-member (NoteFieldRenderArgs::
    draw_pixels_{after,before}_targets) args passed to IsOnScreen()'s
    int parameters -> static_cast<int>, fixing every invocation site.
  * Same float-into-int pattern at the individual DrawBoard() and
    FindFirstDisplayedBeat()/FindLastDisplayedBeat() call sites.
  * DrawBoard(): int - int assigned to a float -> static_cast<float>.
  * CalcPixelsBeforeAndAfterTargets(): (int)(float) round-tripped back
    into a float member -> outer static_cast<float> documents the
    truncate-to-whole-pixels intent.
  * get_returned_column(): lua_tonumber() (double) truncated to an
    index int -> static_cast<int> before the `- 1`.
- TimingSegments.cpp: *Segment::GetValues() push_back(int-or-enum) into
  a std::vector<float> -- static_cast<float> at each push_back.
- RageSurfaceUtils.cpp: std::trunc(float) (still float) assigned to an
  int pixel index; a clamped std::lrint (long) assigned to
  std::uint8_t -- both wrapped in the narrowing cast they already do.
- ScreenOptionsMasterPrefs.cpp: FindClosestEntry<T,U>'s `if constexpr`
  helper lambda wrapped in static_cast<T>(...) once, fixing it for
  every T/U instantiation used in the codebase.

Verified: sm_tests 5966/226, ctest 100%, Release build + --SelfTest
green, with /wd4244/4267 removed AND restored (187 sites remain
elsewhere, so the suppression stays for now).

+++ src/RageSurfaceUtils.cpp
@@ -424,10 +424,10 @@ void RageSurfaceUtils::BlitTransform( const RageSurface *src, RageSurface *dst,
 			 * pixel[1]; 2 indicates 50% pixel[1], 50% pixel[2] (which is clamped
 			 * to pixel[1]). */
 			int src_x[2], src_y[2];
-			src_x[0] = std::trunc(src_xp - 0.5f);
+			src_x[0] = static_cast<int>( std::trunc(src_xp - 0.5f) );
 			src_x[1] = src_x[0] + 1;
 
-			src_y[0] = std::trunc(src_yp - 0.5f);
+			src_y[0] = static_cast<int>( std::trunc(src_yp - 0.5f) );
 			src_y[1] = src_y[0] + 1;
 
 			// Emulate GL_REPEAT.
@@ -865,10 +865,10 @@ RageSurface *RageSurfaceUtils::PalettizeToGrayscale( const RageSurface *src_surf
 		const unsigned int A = (index & Amask) >> Ashift;
 
 		// if only one intensity value, always fullbright
-		const std::uint8_t ScaledI = Ivalues == 1 ? 255 : std::clamp( std::lrint(I * (255.0f / (Ivalues-1))), 0L, 255L );
+		const std::uint8_t ScaledI = static_cast<std::uint8_t>( Ivalues == 1 ? 255 : std::clamp( std::lrint(I * (255.0f / (Ivalues-1))), 0L, 255L ) );
 
 		// if only one alpha value, always opaque
-		const std::uint8_t ScaledA = Avalues == 1 ? 255 : std::clamp( std::lrint(A * (255.0f / (Avalues-1))), 0L, 255L );
+		const std::uint8_t ScaledA = static_cast<std::uint8_t>( Avalues == 1 ? 255 : std::clamp( std::lrint(A * (255.0f / (Avalues-1))), 0L, 255L ) );
 
 		RageSurfaceColor c;
 		c.r = ScaledI;
