+++ src/TimingSegments.cpp
@@ -176,8 +176,8 @@ RString ComboSegment::ToString(int dec) const
 std::vector<float> ComboSegment::GetValues() const
 {
 	std::vector<float> ret;
-	ret.push_back(GetCombo());
-	ret.push_back(GetMissCombo());
+	ret.push_back(static_cast<float>(GetCombo()));
+	ret.push_back(static_cast<float>(GetMissCombo()));
 	return ret;
 }
 
@@ -203,8 +203,8 @@ RString TimeSignatureSegment::ToString(int dec) const
 std::vector<float> TimeSignatureSegment::GetValues() const
 {
 	std::vector<float> ret;
-	ret.push_back(GetNum());
-	ret.push_back(GetDen());
+	ret.push_back(static_cast<float>(GetNum()));
+	ret.push_back(static_cast<float>(GetDen()));
 	return ret;
 }
 
@@ -222,7 +222,7 @@ std::vector<float> SpeedSegment::GetValues() const
 	std::vector<float> ret;
 	ret.push_back(GetRatio());
 	ret.push_back(GetDelay());
-	ret.push_back(GetUnit());
+	ret.push_back(static_cast<float>(GetUnit()));
 	return ret;
 }
